LINE TRACKING Using KY-033 SENSOR MODULE
The KY-033 Line Tracking module is an infrared sensor that detects whether the surface in front of it is reflective or opaque. Sensitivity to ambient light can be adjusted using the knob to achieve a fairly accurate reading.

This sensor is typically used on wheeled robots and can operate at 3.3V to 5V which makes it compatible with a variety of microcontrollers like Arduino, ESP32, Raspberry Pi, ESP8266, Teensy, and others.