KY-025 REED SWITCH MODULE
The KY-025 Reed Switch Module is a small electrical switch operated by an applied magnetic field, commonly used as proximity sensor.

The module has both digital and analog outputs. A trimmer is used to calibrate the sensitivity of the sensor.

In this Arduino sketch we’ll read values from both digital and analog interfaces on the KY-025, you’ll need a magnet to interact with the module.

The digital interface will send a HIGH signal when a magnetic field is detected, turning on the LED on the Arduino (pin 13).

On the other hand, the analog interface will return a HIGH numeric value when there’s no magnetic field present and it’ll drop to zero when a magnet is near.