KY-035 ANALOG HALL MAGNETIC SENSOR MODULE
The KY-035 analog magnetic hall sensor is a module that reacts in the presence of a magnet. It can measure the polarity and relative strength of a magnetic field.

When no magnetic field is detected, the output signal is roughly half of the applied input voltage. When a magnetic field is detected, the output signal goes up or down depending on the polarity and proximity of the magnet.

This module offers a more detailed proximity reading than the KY-003, a digital magnetic sensor that looks very similar to this module. It is also functionally similar to the KY-024, a digital/analog magnetic sensor.